adding mocha unit test suite
