+++ .github/workflows/build.yml
@@ -21,6 +21,7 @@ jobs:
           node-version: ${{ matrix.node-version }}
       - uses: bahmutov/npm-install@v1
       - run: npm run lint
+      - run: npm run test
       - run: npm run dist
       - run: npm run apidocs
       - run: npm run installdocs
