+++ package.json
@@ -31,9 +31,8 @@
     "distdocs": "cd docs && node node_modules/gatsby/cli.js build --prefix-paths --no-uglify",
     "apidocs": "node node_modules/@microsoft/api-extractor/bin/api-extractor run --local && node node_modules/@microsoft/api-documenter/bin/api-documenter markdown --input-folder etc --output-folder built/apidocs",
     "watch": "node node_modules/rollup/dist/bin/rollup -c rollup.config.ts -w",
-    "test": "node node_modules/jest/bin/jest.js",
-    "test:watch": "node node_modules/jest/bin/jest.js --watch",
-    "test:prod": "npm run lint && npm run test --no-cache",
+    "test": "node node_modules/ts-mocha/bin/ts-mocha -p ./tsconfig.spec.json src/**/*.spec.ts",
+    "testwatch": "node node_modules/ts-mocha/bin/ts-mocha -p ./tsconfig.spec.json src/**/*.spec.ts -w --watch-extensions ts",
     "tools": "npx http-server . -c-1",
     "installdocs": "cd docs && npm install --force",
     "installdev": "npm install && npm run installdocs",
@@ -137,11 +136,14 @@
     "@microsoft/api-extractor": "^7.13.1",
     "@semantic-release/exec": "^5.0.0",
     "@semantic-release/git": "^9.0.0",
+    "@types/expect": "^24.3.0",
+    "@types/mocha": "^8.2.1",
     "cli": "^1.0.1",
     "colors": "^1.4.0",
     "cross-env": "^7.0.3",
     "lint-staged": "^10.5.4",
     "lodash.camelcase": "^4.3.0",
+    "mocha": "^8.3.0",
     "prompt": "^1.1.0",
     "replace-in-file": "^6.2.0",
     "rollup": "^2.38.5",
@@ -152,6 +154,7 @@
     "rollup-plugin-typescript2": "^0.29.0",
     "semantic-release": "^17.3.8",
     "shelljs": "^0.8.4",
+    "ts-mocha": "^8.0.0",
     "tslint": "^6",
     "tslint-microsoft-contrib": "^6",
     "typescript": "^4.1.4",
